+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project doesn't care about versioning.
 ### Added
 - Add support for multi-root workspaces, a.k.a. `.code-workspace` files (see [GH-15]).
 
+### Changed
+- Use async IO.
+
 [GH-15]: https://github.com/lunaryorn/gnome-search-providers-vscode/pull/15
 
 ## [1.5.0] – 2021-09-25