+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project doesn't care about versioning.
 
 ## [Unreleased]
 
+## [1.6.0] – 2021-11-27
+
 ### Added
 - Add support for multi-root workspaces, a.k.a. `.code-workspace` files (see [GH-15]).
 
@@ -84,7 +86,8 @@ and this project doesn't care about versioning.
 
 Initial release with support for workspaces of Code - OSS from Arch Linux.
 
-[Unreleased]: https://github.com/lunaryorn/gnome-search-providers-vscode/compare/v1.5.0...HEAD
+[Unreleased]: https://github.com/lunaryorn/gnome-search-providers-vscode/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/lunaryorn/gnome-search-providers-vscode/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/lunaryorn/gnome-search-providers-vscode/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/lunaryorn/gnome-search-providers-vscode/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/lunaryorn/gnome-search-providers-vscode/compare/v1.2.0...v1.3.0